plasma: fixup and cleanup
the default layout is done via the scripting support, that's just
a no go. plasma-default-layoutrc is possible simpler solution with
for the madness that the default plasma settings are done on
initial setup but it most likely some features will have to be
extended, especially the containment size option. hardcoding panel
and default view to always be available by setting virtual desktops
and other misc settings accordingly seems better but that will
probably break more code thus avoiding it for now