kwin: drop wayland support
altought it would be nice to have support for it, wayland is not
yet ready for prime time. it breaks from time to time and with
version 1.6.0 it can cause serious issues (kwin crashes). there is
xwayland and I doubt X11/X.Org/XFree86/whatever will be gone, not
just yet.