kwin: change oxygen frame border to BorderDefault
I do not know the reaseons behind making it default to BorderTiny
but that makes the windows content look like a set toolbar, panel,
window and a statusbar (in most cases) withtout visual indicator
they are part of the same window - without a frame.

also, the indicator for horizontal resizing of the window is now a
bit more noticable for the eye.

the only problem left to adress is probably the spacing of the left
side of the scrollbars which looks kinda too small and may need
adjustment to make windows (as a whole) look a bit more pleasant
to look at.