kdontchangethehostname: fixup, it was utterly broken
it seems the the compiler got confused by the 3-space identation
and due to double space field separation the parsing code was not
working at all. now if kded starts monitoring for hostname change
(the code for that is only a comment) changing the hostname should
not prevent X11 apps from working.