plasma: make plasma imported work with Python 3
the Python byte-compiler does not care much about variables and
tries to byte-compile the Python 2 code from the file leading to
error during build.

with this commit Python 2.7 or newer are requirement but nowdays
distributions are (slowly) moving towards trowing out Python 2
and replacing it with Python 3 so I do not think this is going to
be much of a problem.