kcontrol: remove check if lid is closed
since the randr module methods are not D-Bus exported there is (normally)
no way for a user to interact with the action and trigger display switch,
the system may even be suspended while the lid is closed. there is also the
case where the laptop screen is not used (external monitor is used), lid is
closed, the system is not suspended and a switch has to happen from the
external monitor to the laptop screen because the external monitor is
unplugged. the check is just bogus for many cases

Signed-off-by: Ivailo Monev <xakepa10@gmail.com>