ksmserver: log backtrace in case of crash
if it crashes a startup error will be shown via `xmessage`, there will
be not window manager (most likely) at that point and many things will
not work as expected.

ksmserver cannot be restarted for example, drkonqi is not option either
because attempting to report a bug will open a browser but the backtrace
cannot be copied (the browser window will be on top of any other window
if it launches with no way to switch to the drkonqi window), etc.

in other words - the best thing that can be done is log the backtrace to
the system log (the ksmserver debug area has to be configured for that)

Signed-off-by: Ivailo Monev <xakepa10@gmail.com>