plasma: do not translate the wind directions from IonInterface::getWindDirectionIcon()
fixes weatherstation plasma applet not showing the wind image (wind_arrows)
when the locale is set to be one other than the default (C/en_US). note
that "N/A" is matched as untranslated string (e.g. in
kdeplasma-addons/applets/weather/weatherapplet.cpp) meaning more changes
may be required to fix other minor issues because "N/A" is translated in
several places

Signed-off-by: Ivailo Monev <xakepa10@gmail.com>