generic: apply effect settings from the platform plugin instead of writing to the Katie config
the idea being to not write anything into the Katie config and possibly
even drop support for overriding Katie settings from config.

on a side note font and palette settings are wrapped via classes
(KGlobalSettings and KColorScheme) and hard-wired to configs rather than
being a runtime thing (e.g. getting the font to use for menus via
QApplication::font("QMenu") instead of KGlobalSettings::menuFont()) so lots
to do on that, as for how it will benefit Katana and Katie - less config
parsing possibly, QPalette objects are shared data for example

Signed-off-by: Ivailo Monev <xakepa10@gmail.com>