plasma: reimplement tasks applet
very dull right now but atleast text eliding is done correctly. lots of
plans for it, tasks grouping and sorting is unlikely to be a thing tho.
the people behind the old implementation missing a few basic things -
what things do I want the taskbar to show (e.g. icon and text, icon or text
only) and what happens when a task is interacted with (on
left/right/middle mouse button click - there can be option to move tasks
on middle click for example)

Signed-off-by: Ivailo Monev <xakepa10@gmail.com>